+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Set the process title according to the current application name and task name
 * Ignore some low level env vars in runnable command showed in logs
 * Fix section output to work on Windows
+* Add a `compile` command that puts together a customizable PHP binary with a repacked castor app into one executable file
 
 ## 0.12.1 (2024-02-06)